move big converter methods to model converter component
