+++ src/main/java/org/cyclonedx/maven/ModelConverter.java
@@ -19,7 +19,37 @@
 package org.cyclonedx.maven;
 
 import org.apache.maven.artifact.Artifact;
+import org.apache.maven.project.MavenProject;
+import org.cyclonedx.CycloneDxSchema;
+import org.cyclonedx.model.Component;
+import org.cyclonedx.model.Metadata;
 
+/**
+ * Model converter from Maven concepts (Artifact + MavenProject) to CycloneDX ones
+ * (resp. pURL and Component + Metadata).
+ */
 public interface ModelConverter {
     String generatePackageUrl(Artifact artifact);
+
+    /**
+     * Converts a Maven artifact (dependency or transitive dependency) into a
+     * CycloneDX component.
+     *
+     * @param artifact the artifact to convert
+     * @param schemaVersion the target CycloneDX schema version
+     * @param includeLicenseText should license text be included in bom?
+     * @return a CycloneDX component
+     */
+    Component convert(Artifact artifact, CycloneDxSchema.Version schemaVersion, boolean includeLicenseText);
+
+    /**
+     * Converts a MavenProject into a Metadata object.
+     *
+     * @param project the MavenProject to convert
+     * @param projectType the target CycloneDX component type
+     * @param schemaVersion the target CycloneDX schema version
+     * @param includeLicenseText should license text be included in bom?
+     * @return a CycloneDX Metadata object
+     */
+    Metadata convert(MavenProject project, String projectType, CycloneDxSchema.Version schemaVersion, boolean includeLicenseText);
 }
